refactor(coordinator): 抽出 capsule 子模块
机械式行为保持拆分：从 coordinator.rs 搬出 capsule HUD 呈现层 —— 显示策略、
emit_capsule 状态广播、布局快照/底部居中定位、自动回 Idle 调度，及相关 const/static。
emit_capsule 的 ~30Hz 音频回调线程 run_on_main_thread marshaling 原样保留。
CapsuleLayoutState 经 glob 重导出供 Inner 字段类型可见。
